chore: 분리된 api 파일을 단일 파일로 병합
