+++ src/features/chat/index.ts
@@ -1,19 +1,16 @@
-export { default as FriendList } from "./ui/FriendList";
-export { default as ChatList } from "./ui/ChatList";
-export { default as ChatListItem } from "./ui/ChatListItem";
-export { default as FriendListItem } from "./ui/FriendListItem";
-
-// apis
-export { default as fetchMessagesForChat } from "./apis/fetchMessagesForChat";
-export { default as sendChatMessage } from "./apis/sendChatMessage";
-export { default as fetchUserFriends } from "./apis/fetchUserFriends";
+export { default as FriendList } from './ui/FriendList';
+export { default as ChatList } from './ui/ChatList';
+export { default as ChatListItem } from './ui/ChatListItem';
+export { default as FriendListItem } from './ui/FriendListItem';
 
 // hooks
-export { default as useRealTimeMessages } from "./hooks/useRealTimeMessages";
+export { default as useRealTimeMessages } from './hooks/useRealTimeMessages';
 
 // model
-export { default as useChatsQuery } from "./model/useChatsQuery";
-export { default as useFriendsQuery } from "./model/useFriendsQuery";
-export { default as useMessagesQuery } from "./model/useMessageQuery";
-export { default as useMessageMutation } from "./model/useMessageMutation";
+export { default as useChatsQuery } from './model/useChatsQuery';
+export { default as useFriendsQuery } from './model/useFriendsQuery';
+export { default as useMessagesQuery } from './model/useMessageQuery';
+export { default as useMessageMutation } from './model/useMessageMutation';
 
+// apis
+export * from './apis';
