+++ src/features/chat/ui/ChatList.tsx
@@ -1,12 +1,12 @@
-import { userIdAtom } from "@/entities/auth";
-import { ChatListItem, useChatsQuery } from "@/features/chat";
-import * as stylex from "@stylexjs/stylex";
-import { useAtom } from "jotai";
-import { memo } from "react";
+import { userIdAtom } from '@/entities/auth';
+import { ChatListItem, useChatsQuery } from '@/features/chat';
+import * as stylex from '@stylexjs/stylex';
+import { useAtom } from 'jotai';
+import { memo } from 'react';
 
 const ChatList = () => {
   const [userId] = useAtom(userIdAtom);
-  const { data, isLoading } = useChatsQuery(userId);
+  const { data, isLoading } = useChatsQuery(userId!);
 
   if (isLoading) return <div>Loading...</div>;
 
@@ -23,8 +23,8 @@ export default memo(ChatList);
 
 const styles = stylex.create({
   list: {
-    display: "flex",
-    flexDirection: "column",
-    gap: "20px",
+    display: 'flex',
+    flexDirection: 'column',
+    gap: '20px',
   },
 });
